pyproject.toml
[project]
name = "agvmon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "pika>=1.3.2",
    # "pillow>=12.1.0",
    "pyzmq>=27.1.0",
]
